Added some functions to dmeditor.py for error underlining.
Call 'addError(line number)' on the dmeditor showing the error in question when you want to add an error. Call 'clearErrors()' when you want to get rid of all the error indicators.

If people change any of the error-indicated text, it loses the underline, but it moves around appropriately if they add or delete characters before it.

git-svn-id: http://dmide.googlecode.com/svn@88 a812e4ac-3740-0410-bb29-cddadda57e64